+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Generate flags moved to root command: `llm-imager -p "..." -o "..."` now works without `generate` subcommand
+
 ## [0.1.4] - 2026-02-22
 
 ### Added